Combine verbosity-related bool flags into 1 int flag
Previously, there were 3 different bool flags related to verbosity.
Having many member variables for values that should be mutually
exclusive led to a bug where contradicting flags could be set. This
caused unexpected behavior with regards to logging.

This commit combines the bool flags into 1 int flag. As a result,
verbosity level is guaranteed to be mutually exclusive. This commit also
adds validation of the verbosity flags and exits with an error msg.

Signed-off-by: Myung-jong Kim <mjkim610@gmail.com>