Combine verbosity-related bool flags into single enum member variable (#148)
* Combine verbosity-related bool flags into 1 int flag

Previously, there were 3 different bool flags related to verbosity.
Having many member variables for values that should be mutually
exclusive led to a bug where contradicting flags could be set. This
caused unexpected behavior with regards to logging.

This commit combines the bool flags into 1 int flag. As a result,
verbosity level is guaranteed to be mutually exclusive. This commit also
adds validation of the verbosity flags and exits with an error msg.

Signed-off-by: Myung-jong Kim <mjkim610@gmail.com>

* Create Verbosity enum

This commit defines a new enum Verbosity. The log level flag
Verbosity is converted from int to Verbosity. This has the advantage of
clearly defining the human-understandable string values to each choice.
It also prevents bugs where an unexpected int value is assigned.

Signed-off-by: Myung-jong Kim <mjkim610@gmail.com>

Signed-off-by: Myung-jong Kim <mjkim610@gmail.com>